+++ frame_source/source_imagefile.cpp
