+++ frame_source/source_imagefile.h
